update TextInput stories to latest recommendations
